+++ src/main.ts
@@ -3,7 +3,7 @@ import pinia from "@/stores/pinia";
 import App from "@/App.vue";
 import router from "@/router";
 import config from "@/config";
-import { sidenavConfigKey, enabledPrezsConfigKey, apiBaseUrlConfigKey, configKey } from "@/types";
+import { sidenavConfigKey, enabledPrezsConfigKey, apiBaseUrlConfigKey, mapConfigKey } from "@/types";
 
 import VueGoogleMaps from '@fawmi/vue-google-maps'
 
@@ -11,19 +11,18 @@ import "@/assets/sass/main.scss";
 
 const app = createApp(App);
 
-
 app.provide(sidenavConfigKey, config.sidenav === "true");
 app.provide(enabledPrezsConfigKey, config.enabledPrezs.split(","));
 app.provide(apiBaseUrlConfigKey, config.apiBaseUrl);
+app.provide(mapConfigKey, config.map);
 
 app.use(pinia);
 app.use(router);
 app.use(VueGoogleMaps, {
     load: {
-        key: config.mapSettingsApiKey,
+        key: config.map.settings.apiKey,
         libraries: "drawing"
     },
 })
-
 
 app.mount("#app");